+++ src/App.tsx
@@ -1,9 +1,11 @@
-import { Button } from "./components/ui/button";
+import { RouterProvider } from 'react-router-dom';
+import router from './router';
+
 function App() {
+  
   return (
     <>
-      <h1 className="text-3xl font-bold underline">Hello world!</h1>
-      <Button>Button</Button>
+      <RouterProvider router={router}/>
     </>
   );
 }