+++ CHANGELOG.md
@@ -1,3 +1,14 @@
+## 0.1.1 ( 2025-01-01 )
+
+### feat
+
+- Add todolist testing
+- Add router table
+
+### chore
+
+- Add react-router plugin
+
 ## 0.1.0 ( 2024-12-04 )
 
 ### chore